:recycle: refactor: move tuto datasource to `src`
